@@ -1,9 +1,17 @@
+import type { Metadata } from "next";
+import { Vazirmatn } from "next/font/google";
 import "./globals.css";
-import { ThemeProvider } from "../contexts/ThemeContext";
+import { ThemeProvider } from "@/contexts/ThemeContext";
 
-export const metadata = {
-  title: "NOWEX Admin Panel",
-  description: "پنل مدیریت نواکس",
+const vazir = Vazirmatn({
+  subsets: ["arabic"],
+  weight: ["300", "400", "500", "700", "900"],
+  variable: "--font-vazir",
+});
+
+export const metadata: Metadata = {
+  title: "پنل مدیریت NOWEX",
+  description: "پنل مدیریتی پلتفرم NOWEX",
 };
 
 export default function RootLayout({
@@ -13,7 +21,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="fa" dir="rtl">
-      <body className="bg-gray-50 dark:bg-gray-900 text-gray-900 dark:text-gray-100">
+      <body className={`${vazir.variable} font-sans antialiased`}>
         <ThemeProvider>{children}</ThemeProvider>
       </body>
     </html>
