@@ -1,25 +1,50 @@
 "use client";
 
-import { useTheme } from "../../contexts/ThemeContext";
-import { SunIcon, MoonIcon } from "@heroicons/react/24/outline";
+import { useTheme } from "@/contexts/ThemeContext";
 
 export default function Header() {
-  const { theme, toggleTheme } = useTheme();
+  const { colors, toggleTheme, theme } = useTheme();
 
   return (
-    <header className="w-full h-16 bg-gray-100 dark:bg-gray-800 flex items-center justify-between px-6 shadow">
-      <h1 className="text-lg font-semibold">پنل مدیریت نواکس</h1>
+    <header
+      className="w-full flex items-center justify-between px-6 py-4 shadow-md"
+      style={{
+        background: `linear-gradient(90deg, ${colors.headerBg1}, ${colors.headerBg2})`,
+        color: colors.headerText,
+        borderBottom: `1px solid ${colors.border1}`,
+      }}
+    >
+      {/* Title */}
+      <h1 className="text-xl font-semibold tracking-wide">
+        NOWEX Admin Panel
+      </h1>
 
-      <button
-        onClick={toggleTheme}
-        className="p-2 rounded-full bg-gray-200 dark:bg-gray-700 hover:bg-gray-300 dark:hover:bg-gray-600 transition"
-      >
-        {theme === "light" ? (
-          <MoonIcon className="w-6 h-6" />
-        ) : (
-          <SunIcon className="w-6 h-6" />
-        )}
-      </button>
+      {/* Right Section */}
+      <div className="flex items-center gap-4">
+        {/* Theme Toggle */}
+        <button
+          onClick={toggleTheme}
+          className="px-4 py-2 rounded-md transition-all"
+          style={{
+            background: colors.button1,
+            color: colors.headerText,
+            border: `1px solid ${colors.border1}`,
+          }}
+        >
+          {theme === "dark" ? "Light Mode" : "Dark Mode"}
+        </button>
+
+        {/* Admin Profile */}
+        <div
+          className="px-4 py-2 rounded-md"
+          style={{
+            background: colors.cardBg1,
+            border: `1px solid ${colors.border1}`,
+          }}
+        >
+          Admin
+        </div>
+      </div>
     </header>
   );
 }
