@@ -1,3 +1,6 @@
-﻿export { Button } from "./Button"
+﻿"use client";
+export * from "./theme/ThemeContext";
+export { default as Logo } from "./components/Logo";
+export { Button } from "./Button"
 export { Input } from "./Input"
 export { Card } from "./Card"
